requirements.txt
Flask==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1

sqlparse==0.4.1
urllib3==1.26.3
Werkzeug==1.0.1

pip~=21.1.3
mysql-connector-python~=8.0.26
SQLAlchemy~=1.4.21